requirements.txt
Abg==2.3.1
git+https://github.com/KurimuzonAkuma/pyrogram@dev
pymongo
motor
requests
dnspython
tgcrypto
heroku3
asyncio
python-dotenv
lexica-api
deep-translator
pycgi
apscheduler
Flask
uvloop
psutil
langdetect
MukeshAPI
speedtest-cli
aiohttp
gitpython
